requirements.txt
llama-index==0.4.33
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-decouple==3.8
langchain~=0.0.128
pytest==7.2.2 
PyPDF2==3.0.1
boto3==1.26.104 